Updated Background, added textures, and imported texture files
This update added a bunch of the texture files to the game, changed the
background java file to do a switch/case to be able to import the
backgrounds needed quickly and with as little change as possible. Change
will need to be made to keep track of world and feed that into the
background class. Currently hard coded to be 1. Poison was added as a
character with movement disabled in order to be able to check for
collisions and do damage to the player. The breakable walls were added
as a extended blocks. Challenging and Super Tough enemies are now
associated with textures as well.

To Do:
Add the breakable walls to the interacts list for the appropriate
weapons.
Add Bridge and Abyss to possible floors.
Add list to figure out what world the game is on and send that into
Background to the switch function.